refactor

+++ src/domain/period.rs
@@ -8,8 +8,8 @@ use serde::{Serialize, Deserialize};
 pub struct Period {
     pub id: i32,
     pub name: String,
-    pub year_start: i32,
-    pub year_end: Option<i32>,
+    pub year_start: i16,
+    pub year_end: Option<i16>,
     pub slug: String, // Unique period readable text id, to be used in URLs.
     pub composers: Vec<Composer>,
 }
